perf: flatten asset_holdings to one composite-key map; exact u128 quantities
asset_holdings was a deeply nested imbl map
(OrdMap<AssetKey, OrdMap<policy, OrdMap<name, Qty>>>). With ~10M tiny
inner sub-maps, each pre-allocating a fixed-capacity ~2.5KB imbl chunk
that sits nearly empty, the map ballooned to ~22GB on mainnet and the
cold-start reset was OOM-killed.

Flatten to a single OrdMap<HeldKey, Qty> keyed by
(AssetKey, policy(28)++name). One large map fills its chunks (~234k full
chunks vs ~10M near-empty), dropping asset_holdings to ~3-4GB. Per-address
and per-credential access become contiguous prefix range scans over the
sorted composite key (reusing the existing cred_range pattern); a block
only mutates moved tokens, so point ops stay O(log n) — whale-safe — and
the live tile diffs walk prev.diff(curr), bounded by the block's changes.

Also preserve exact asset quantities: Qty is u128 with variable-length
serde (u64 when it fits, else a (lo,hi) pair), and the holdings build
sums via ::text to avoid the bigint overflow that previously crashed
reset. Bump SNAPSHOT_FORMAT 3 -> 4 (persisted shape changed).

Adds a memory watchdog thread and periodic log_sizes for RSS visibility.

+++ server/src/mempool.rs
@@ -111,7 +111,7 @@ pub async fn extract_tx(
                             Some(AssetInfo {
                                 fingerprint,
                                 name,
-                                quantity: format_quantity(raw, decimals),
+                                quantity: format_quantity(raw as u128, decimals),
                                 tks,
                                 tk: None,
                                 size: 0,
